+++ Ergo/Runtime/Solutions/ISolutionEnumerable.cs
@@ -1,0 +1,3 @@
+﻿namespace Ergo.Runtime;
+
+public interface ISolutionEnumerable : IReadOnlyList<Solution> { }
