+++ ErgoVSIX/ErgoLS/ergo/stdlib/Ergo.runtimeconfig.json
@@ -6,6 +6,7 @@
       "version": "8.0.0-rc.1.23419.4"
     },
     "configProperties": {
+      "System.Reflection.Metadata.MetadataUpdater.IsSupported": false,
       "System.Runtime.Serialization.EnableUnsafeBinaryFormatterSerialization": false
     }
   }
